[refactor] custom hook 제작 modal/match #775
